+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to Git Syncher will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.1] - 2026-08-18
+
+### Changed
+
+- Adding a folder that is already on the dashboard for that branch **updates** that card instead of blocking with an error
+
+### Fixed
+
+- Account picker fills Git username and email as soon as you select an account
+
 ## [1.12.0] - 2026-08-18
 
 ### Changed